Summarise the characteristics of the base population of a cdm object.

Summarise the characteristics of the base population of a cdm object.

Usage

summarisePopulationCharacteristics(
  cdm,
  studyPeriod = c(NA, NA),
  sex = FALSE,
  ageGroup = NULL
)

Arguments

cdm

A cdm object.

studyPeriod

Dates to trim the observation period. If NA, start_observation_period and/or end_observation_period are used.

sex

Boolean variable. Whether to stratify the results by sex.

ageGroup

List of age groups to stratify by at index date. Set to NULL if no stratification is needed.

Value

A summarised_result object.

Examples

# \donttest{
library(dplyr)
library(CDMConnector)
library(DBI)
library(duckdb)
library(OmopSketch)

# Connect to Eunomia database
if (Sys.getenv("EUNOMIA_DATA_FOLDER") == "") Sys.setenv("EUNOMIA_DATA_FOLDER" = tempdir())
if (!dir.exists(Sys.getenv("EUNOMIA_DATA_FOLDER"))) dir.create(Sys.getenv("EUNOMIA_DATA_FOLDER"))
if (!eunomia_is_available()) downloadEunomiaData()
con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdmFromCon(
con = con, cdmSchema = "main", writeSchema = "main"
)

# Run summarise clinical tables
summarisedPopulation <- summarisePopulationCharacteristics(cdm = cdm,
                                                          studyPeriod = c("2010-01-01",NA),
                                                          sex = TRUE,
                                                          ageGroup = NULL
                                                          )
#> Warning: ! 1 casted column in og_001_1726476626 (cohort_set) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 1 column in og_001_1726476626 do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ℹ Building new trimmed cohort
#> Warning: ! 1 column in tmp_001_og_003_1726476626 do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> Creating initial cohort
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ✔ Cohort trimmed
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
summarisedPopulation |> print()
#> # A tibble: 122 × 13
#>    result_id cdm_name            group_name group_level strata_name strata_level
#>        <int> <chr>               <chr>      <chr>       <chr>       <chr>       
#>  1         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  2         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  3         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  4         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  5         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  6         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  7         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  8         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#>  9         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#> 10         1 Synthea synthetic … cohort_na… demographi… overall     overall     
#> # ℹ 112 more rows
#> # ℹ 7 more variables: variable_name <chr>, variable_level <chr>,
#> #   estimate_name <chr>, estimate_type <chr>, estimate_value <chr>,
#> #   additional_name <chr>, additional_level <chr>
PatientProfiles::mockDisconnect(cdm = cdm)
# Change to mock data whenever possible
# }