Summarise table quality issues in an OMOP table using a specific time interval.

This function assesses potential data quality issues in a given OMOP table, such as records with an end date before the start date or a start date before the person's birthdate.

Usage

summariseTableQuality(
  cdm,
  omopTableName,
  interval = "overall",
  sex = FALSE,
  ageGroup = NULL,
  sample = NULL,
  dateRange = NULL,
  endBeforeStart = TRUE,
  startBeforeBirth = TRUE
)

Arguments

cdm

A cdm_reference object.

omopTableName

A character vector of OMOP table names from the CDM.

interval

Time interval to stratify by. It can either be "years", "quarters", "months" or "overall".

sex

Logical. Should results be stratified by sex? Default is FALSE.

ageGroup

A list of age groups to stratify results by. Default is NULL.

sample

An integer specifying the number of records to sample from each table. If NULL, no sampling is done.

dateRange

A vector of two dates defining the desired study period. Only the start_date column of the OMOP table is checked to ensure it falls within this range. If dateRange is NULL, no restriction is applied.

endBeforeStart

Logical. Whether the function should check for records where the end date is before the start date. Default is TRUE.

startBeforeBirth

Logical. Whether the function should check for records where the start date is before the patient's birthdate. Default is TRUE.

Value

A summarised_result object with counts and percentages of potential quality issues, stratified as specified.

Examples

# \donttest{
library(dplyr)

cdm <- mockOmopSketch()

qualitySummary <- summariseTableQuality(
  cdm = cdm,
  omopTableName = c("condition_occurrence", "drug_exposure"),
  interval = "years",
  sex = TRUE,
  ageGroup = list("<=20" = c(0, 20), ">20" = c(21, Inf)),
  endBeforeStart = TRUE,
  startBeforeBirth = TRUE
)

qualitySummary |> glimpse()
#> Rows: 8
#> Columns: 13
#> $ result_id        <int> 1, 1, 1, 1, 1, 1, 1, 1
#> $ cdm_name         <chr> "mockOmopSketch", "mockOmopSketch", "mockOmopSketch",…
#> $ group_name       <chr> "omop_table", "omop_table", "omop_table", "omop_table…
#> $ group_level      <chr> "condition_occurrence", "condition_occurrence", "cond…
#> $ strata_name      <chr> "overall", "overall", "overall", "overall", "overall"…
#> $ strata_level     <chr> "overall", "overall", "overall", "overall", "overall"…
#> $ variable_name    <chr> "Records with end date before start date", "Records w…
#> $ variable_level   <chr> NA, NA, NA, NA, NA, NA, NA, NA
#> $ estimate_name    <chr> "count", "count", "percentage", "percentage", "count"…
#> $ estimate_type    <chr> "integer", "integer", "percentage", "percentage", "in…
#> $ estimate_value   <chr> "0", "0", "0.00", "0.00", "0", "0", "0.00", "0.00"
#> $ additional_name  <chr> "overall", "overall", "overall", "overall", "overall"…
#> $ additional_level <chr> "overall", "overall", "overall", "overall", "overall"…

PatientProfiles::mockDisconnect(cdm = cdm)
# }