Create a visual table from a summarise_omop_snapshot result

Create a visual table from a summarise_omop_snapshot result

Usage

tableOmopSnapshot(result, type = NULL, style = NULL)

Arguments

result

A summarised_result object (output of summariseOmopSnapshot() ).

type

Character string specifying the desired output table format. See visOmopResults::tableType() for supported table types. If type = NULL, global options (set via visOmopResults::setGlobalTableOptions()) will be used if available; otherwise, a default 'gt' table is created.

style

Defines the visual formatting of the table. This argument can be provided in one of the following ways:

Pre-defined style: Use the name of a built-in style (e.g., "darwin"). See visOmopResults::tableStyle() for available options.

YAML file path: Provide the path to an existing .yml file defining a new style.

List of custome R code: Supply a block of custom R code or a named list describing styles for each table section. This code must be specific to the selected table type.

If style = NULL, the function will use global options (see visOmopResults::setGlobalTableOptions()) or an existing ⁠_brand.yml⁠ file (if found); otherwise, the default style is applied.

Value

A formatted table visualisation.

Examples

# \donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")
#> ℹ Reading GiBleed tables.
#> ℹ Adding drug_strength table.
#> ℹ Creating local <cdm_reference> object.
#> ℹ Inserting <cdm_reference> into duckdb.

result <- summariseOmopSnapshot(cdm = cdm)

tableOmopSnapshot(result = result)

  
Snapshot of the cdm GiBleed
  
Estimate
      	
        
Database name

      
    

GiBleed
    

General
    

Snapshot date
	2025-11-18

Person count
	2,694

Vocabulary version
	v5.0 18-JAN-19

Cdm
    

Source name
	Synthea synthetic health database

Version
	5.3

Holder name
	OHDSI Community

Release date
	2019-05-25

Description
	SyntheaTM is a Synthetic Patient Population Simulator. The goal is to output synthetic, realistic (but not real), patient data and associated health records in a variety of formats.

Documentation reference
	https://synthetichealth.github.io/synthea/

Observation period
    

N
	2,694

Start date
	1908-09-22

End date
	2019-07-03

Cdm source
    

Type
	duckdb

Package
	CDMConnector

Write schema
	results

Write prefix
	test_

cdmDisconnect(cdm = cdm)
# }