Create a visual table from a summarise_population_characteristics result.

Create a visual table from a summarise_population_characteristics result.

Usage

tablePopulationCharacteristics(result, type = "gt")

Arguments

result

Output from summarisePopulationCharacteristics().

type

Type of formatting output table, either "gt" or "flextable".

Value

A gt or flextable object with the summarised data.

Examples

# \donttest{
cdm <- mockOmopSketch()

summarisedPopulation <- summarisePopulationCharacteristics(
  cdm = cdm,
  studyPeriod = c("2010-01-01", NA),
  sex = TRUE,
  ageGroup = list("<=60" = c(0, 60), ">60" = c(61, Inf))
)
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ℹ Building new trimmed cohort
#> Creating initial cohort
#> ✔ Cohort trimmed
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! The following column type were changed:
#> • variable_name: from integer to character

summarisedPopulation |>
  suppress(minCellCount = 5) |>
  tablePopulationCharacteristics()

  

Variable name
      	Variable level
      	Estimate name
      	
        
Database name

      
    

mockOmopSketch
    

overall; overall
    

Number records
	-
	N
	61

Number subjects
	-
	N
	61

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2013-07-09]

	
	Range
	2010-01-01 to 2019-12-13

Cohort end date
	-
	Median [Q25 - Q75]
	2016-09-02 [2012-08-22 - 2018-06-06]

	
	Range
	2010-02-04 to 2019-12-22

Age at start
	-
	Median [Q25 - Q75]
	31 [22 - 46]

	
	Mean (SD)
	34.46 (15.25)

	
	Range
	10 to 64

Age at end
	-
	Median [Q25 - Q75]
	34 [27 - 50]

	
	Range
	12 to 66

Sex
	Female
	N%
	33 (54.10)

	Male
	N%
	28 (45.90)

Prior observation
	-
	Median [Q25 - Q75]
	1,555 [0 - 5,254]

	
	Mean (SD)
	3,524.92 (4,498.94)

	
	Range
	0 to 14,727

Future observation
	-
	Median [Q25 - Q75]
	1,255 [641 - 2,351]

	
	Mean (SD)
	1,422.98 (1,028.71)

	
	Range
	9 to 3,602

Days in cohort
	-
	Median [Q25 - Q75]
	1,256 [642 - 2,352]

	
	Mean (SD)
	1,423.98 (1,028.71)

	
	Range
	10 to 3,603

<=60; overall
    

Number records
	-
	N
	59

Number subjects
	-
	N
	59

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2012-03-28]

	
	Range
	2010-01-01 to 2019-12-13

Cohort end date
	-
	Median [Q25 - Q75]
	2016-08-23 [2012-08-19 - 2018-05-19]

	
	Range
	2010-02-04 to 2019-12-22

Age at start
	-
	Median [Q25 - Q75]
	30 [22 - 44]

	
	Mean (SD)
	33.51 (14.58)

	
	Range
	10 to 60

Age at end
	-
	Median [Q25 - Q75]
	33 [27 - 48]

	
	Range
	12 to 66

Sex
	Female
	N%
	32 (54.24)

	Male
	N%
	27 (45.76)

Prior observation
	-
	Median [Q25 - Q75]
	1,802 [0 - 5,705]

	
	Mean (SD)
	3,644.41 (4,527.19)

	
	Range
	0 to 14,727

Future observation
	-
	Median [Q25 - Q75]
	1,293 [650 - 2,394]

	
	Mean (SD)
	1,449.51 (1,035.31)

	
	Range
	9 to 3,602

Days in cohort
	-
	Median [Q25 - Q75]
	1,294 [650 - 2,394]

	
	Mean (SD)
	1,450.51 (1,035.31)

	
	Range
	10 to 3,603

>60; overall
    

Number records
	-
	N
	<5

Number subjects
	-
	N
	<5

Cohort start date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Cohort end date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Age at start
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Age at end
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Sex
	Female
	N%
	<5 (<5)

	Male
	N%
	<5 (<5)

Prior observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Future observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Days in cohort
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

overall; Female
    

Number records
	-
	N
	33

Number subjects
	-
	N
	33

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2013-11-21]

	
	Range
	2010-01-01 to 2019-12-13

Cohort end date
	-
	Median [Q25 - Q75]
	2017-12-21 [2013-10-28 - 2018-11-03]

	
	Range
	2010-02-04 to 2019-12-22

Age at start
	-
	Median [Q25 - Q75]
	27 [21 - 46]

	
	Mean (SD)
	32.79 (15.86)

	
	Range
	10 to 61

Age at end
	-
	Median [Q25 - Q75]
	32 [23 - 50]

	
	Range
	12 to 66

Sex
	Female
	N%
	33 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	1,859 [0 - 5,254]

	
	Mean (SD)
	3,560.55 (4,572.62)

	
	Range
	0 to 14,727

Future observation
	-
	Median [Q25 - Q75]
	1,396 [564 - 2,511]

	
	Mean (SD)
	1,479.61 (1,125.93)

	
	Range
	9 to 3,602

Days in cohort
	-
	Median [Q25 - Q75]
	1,397 [565 - 2,512]

	
	Mean (SD)
	1,480.61 (1,125.93)

	
	Range
	10 to 3,603

overall; Male
    

Number records
	-
	N
	28

Number subjects
	-
	N
	28

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2010-10-07]

	
	Range
	2010-01-01 to 2019-08-03

Cohort end date
	-
	Median [Q25 - Q75]
	2015-06-18 [2012-07-10 - 2017-12-04]

	
	Range
	2011-07-18 to 2019-12-07

Age at start
	-
	Median [Q25 - Q75]
	32 [28 - 46]

	
	Mean (SD)
	36.43 (14.55)

	
	Range
	12 to 64

Age at end
	-
	Median [Q25 - Q75]
	38 [30 - 48]

	
	Range
	17 to 66

Sex
	Male
	N%
	28 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	1,476 [0 - 5,056]

	
	Mean (SD)
	3,482.93 (4,493.83)

	
	Range
	0 to 14,047

Future observation
	-
	Median [Q25 - Q75]
	987 [687 - 2,135]

	
	Mean (SD)
	1,356.25 (916.86)

	
	Range
	15 to 3,078

Days in cohort
	-
	Median [Q25 - Q75]
	988 [688 - 2,136]

	
	Mean (SD)
	1,357.25 (916.86)

	
	Range
	16 to 3,079

<=60; Female
    

Number records
	-
	N
	32

Number subjects
	-
	N
	32

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2013-08-12]

	
	Range
	2010-01-01 to 2019-12-13

Cohort end date
	-
	Median [Q25 - Q75]
	2017-10-23 [2013-09-23 - 2018-09-18]

	
	Range
	2010-02-04 to 2019-12-22

Age at start
	-
	Median [Q25 - Q75]
	26 [21 - 42]

	
	Mean (SD)
	31.91 (15.27)

	
	Range
	10 to 60

Age at end
	-
	Median [Q25 - Q75]
	32 [23 - 49]

	
	Range
	12 to 66

Sex
	Female
	N%
	32 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	2,211 [0 - 5,617]

	
	Mean (SD)
	3,671.81 (4,600.18)

	
	Range
	0 to 14,727

Future observation
	-
	Median [Q25 - Q75]
	1,452 [467 - 2,520]

	
	Mean (SD)
	1,500.41 (1,137.49)

	
	Range
	9 to 3,602

Days in cohort
	-
	Median [Q25 - Q75]
	1,452 [468 - 2,521]

	
	Mean (SD)
	1,501.41 (1,137.49)

	
	Range
	10 to 3,603

<=60; Male
    

Number records
	-
	N
	27

Number subjects
	-
	N
	27

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2010-06-07]

	
	Range
	2010-01-01 to 2019-08-03

Cohort end date
	-
	Median [Q25 - Q75]
	2015-02-11 [2012-06-21 - 2017-09-06]

	
	Range
	2011-07-18 to 2019-12-07

Age at start
	-
	Median [Q25 - Q75]
	32 [28 - 44]

	
	Mean (SD)
	35.41 (13.76)

	
	Range
	12 to 58

Age at end
	-
	Median [Q25 - Q75]
	37 [30 - 46]

	
	Range
	17 to 66

Sex
	Male
	N%
	27 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	1,555 [0 - 5,423]

	
	Mean (SD)
	3,611.93 (4,526.30)

	
	Range
	0 to 14,047

Future observation
	-
	Median [Q25 - Q75]
	1,016 [713 - 2,149]

	
	Mean (SD)
	1,389.19 (917.29)

	
	Range
	15 to 3,078

Days in cohort
	-
	Median [Q25 - Q75]
	1,017 [714 - 2,150]

	
	Mean (SD)
	1,390.19 (917.29)

	
	Range
	16 to 3,079

>60; Female
    

Number records
	-
	N
	<5

Number subjects
	-
	N
	<5

Cohort start date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Cohort end date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Age at start
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Age at end
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Sex
	Female
	N%
	<5 (<5)

Prior observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Future observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Days in cohort
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

>60; Male
    

Number records
	-
	N
	<5

Number subjects
	-
	N
	<5

Cohort start date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Cohort end date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Age at start
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Age at end
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Sex
	Male
	N%
	<5 (<5)

Prior observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Future observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Days in cohort
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

PatientProfiles::mockDisconnect(cdm = cdm)
# }