Summarise Database Characteristics for OMOP CDM

Summarise Database Characteristics for OMOP CDM

Usage

databaseCharacteristics(
  cdm,
  omopTableName = c("person", "observation_period", "visit_occurrence",
    "condition_occurrence", "drug_exposure", "procedure_occurrence", "device_exposure",
    "measurement", "observation", "death"),
  sex = FALSE,
  ageGroup = NULL,
  dateRange = NULL,
  interval = "overall",
  conceptIdCount = FALSE,
  ...
)

Arguments

cdm

A cdm_reference object representing the Common Data Model (CDM) reference.

omopTableName

A character vector specifying the OMOP tables from the CDM to include in the analysis. If "person" is present, it will only be used for missing value summarisation.

sex

Logical; whether to stratify results by sex (TRUE) or not (FALSE).

ageGroup

A list of age groups to stratify the results by. Each element represents a specific age range.

dateRange

A vector of two dates defining the desired study period. Only the start_date column of the OMOP table is checked to ensure it falls within this range. If dateRange is NULL, no restriction is applied.

interval

Time interval to stratify by. It can either be "years", "quarters", "months" or "overall".

conceptIdCount

Logical; whether to summarise concept ID counts (TRUE) or not (FALSE).

...

additional arguments passed to the OmopSketch functions that are used internally.

Value

A summarised_result object containing the results of the characterisation.

Examples

# \donttest{
cdm <- mockOmopSketch(numberIndividuals = 100)

result <- databaseCharacteristics(cdm)
#> The characterisation will focus on the following OMOP tables: person,
#> observation_period, visit_occurrence, condition_occurrence, drug_exposure,
#> procedure_occurrence, device_exposure, measurement, observation, and death
#> → Getting cdm snapshot
#> Warning: Vocabulary version in cdm_source (NA) doesn't match the one in the vocabulary
#> table (mock)
#> → Getting population characteristics
#> ℹ Building new trimmed cohort
#> Creating initial cohort
#> ✔ Cohort trimmed
#> ℹ adding demographics columns
#> ℹ summarising data
#> ℹ summarising cohort general_population
#> ✔ summariseCharacteristics finished!
#> → Summarising missing data
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: These columns contain missing values, which are not permitted:
#> "race_concept_id" and "ethnicity_concept_id"
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: These columns contain missing values, which are not permitted:
#> "period_type_concept_id"
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: device_exposureomop table is empty.
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> Warning: There was 1 warning in `dplyr::filter()`.
#> ℹ In argument: `as.integer(.data$estimate_value) > 0`.
#> Caused by warning:
#> ! NAs introduced by coercion
#> ! 56 duplicated rows eliminated.
#> → Summarising clinical records
#> ℹ Adding variables of interest to observation_period.
#> ℹ Summarising records per person in observation_period.
#> ℹ Summarising observation_period: `in_observation` and `type_concept`.
#> ℹ Adding variables of interest to visit_occurrence.
#> ℹ Summarising records per person in visit_occurrence.
#> ℹ Summarising visit_occurrence: `in_observation`, `standard_concept`,
#>   `source_vocabulary`, `domain_id`, and `type_concept`.
#> ℹ Adding variables of interest to condition_occurrence.
#> ℹ Summarising records per person in condition_occurrence.
#> ℹ Summarising condition_occurrence: `in_observation`, `standard_concept`,
#>   `source_vocabulary`, `domain_id`, and `type_concept`.
#> ℹ Adding variables of interest to drug_exposure.
#> ℹ Summarising records per person in drug_exposure.
#> ℹ Summarising drug_exposure: `in_observation`, `standard_concept`,
#>   `source_vocabulary`, `domain_id`, and `type_concept`.
#> ℹ Adding variables of interest to procedure_occurrence.
#> ℹ Summarising records per person in procedure_occurrence.
#> ℹ Summarising procedure_occurrence: `in_observation`, `standard_concept`,
#>   `source_vocabulary`, `domain_id`, and `type_concept`.
#> Warning: device_exposure is empty.
#> ℹ Adding variables of interest to measurement.
#> ℹ Summarising records per person in measurement.
#> ℹ Summarising measurement: `in_observation`, `standard_concept`,
#>   `source_vocabulary`, `domain_id`, and `type_concept`.
#> ℹ Adding variables of interest to observation.
#> ℹ Summarising records per person in observation.
#> ℹ Summarising observation: `in_observation`, `standard_concept`,
#>   `source_vocabulary`, `domain_id`, and `type_concept`.
#> ℹ Adding variables of interest to death.
#> ℹ Summarising records per person in death.
#> ℹ Summarising death: `in_observation`, `standard_concept`, `source_vocabulary`,
#>   `domain_id`, and `type_concept`.
#> → Summarising record counts
#> Warning: device_exposure omop table is empty after application of date range.
#> → Summarising in observation records, subjects, person-days, age and sex
#> Warning: Missing values are always removed in SQL aggregation functions.
#> Use `na.rm = TRUE` to silence this warning
#> This warning is displayed once every 8 hours.
#> → Summarising observation period
#> ☺ Database characterisation finished. Code ran in 0 min and 43 sec
#> ℹ 1 table created: "og_005_1747318112".

PatientProfiles::mockDisconnect(cdm)
# }