Generate an interactive Shiny application that visualises the results obtained from the databaseCharacteristics() function

Generate an interactive Shiny application that visualises the results obtained from the databaseCharacteristics() function

Usage

shinyCharacteristics(
  result,
  directory,
  background = TRUE,
  title = "Database characterisation",
  logo = "ohdsi",
  theme = "scarlet"
)

Arguments

result

A summarised_result object (output of databaseCharacteristics()).

directory

A character string specifying the directory where the application will be saved.

background

Background panel for the Shiny app.

If set to TRUE (default), a standard background panel with a general description will be included.

If set to FALSE, no background panel will be displayed.

If it is a path (e.g., "path/to/file.md") tha file will be used as background panel of your shiny App.

title

Title of the shiny. Default is "Characterisation".

logo

Name of a logo or path to a logo. If NULL no logo is included. Only svg format allowed for the moment.

theme

A character string specifying the theme for the Shiny application. It can be any of the OmopViewer supported themes.

Value

This function invisibly returns NULL and generates a static Shiny app in the specified directory.

Examples

if (FALSE) { # \dontrun{
library(OmopSketch)
library(omock)
library(here)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

res <- databaseCharacteristics(cdm = cdm)

shinyCharacteristics(result = res, directory = here())

cdmDisconnect(cdm = cdm)
} # }