Summarise an omop_table from a cdm_reference object. You will obtain information related to the number of records, number of subjects, whether the records are in observation, number of present domains and number of present concepts.

Summarise an omop_table from a cdm_reference object. You will obtain information related to the number of records, number of subjects, whether the records are in observation, number of present domains and number of present concepts.

Usage

summariseClinicalRecords(
  omopTable,
  recordsPerPerson = c("mean", "sd", "median", "q25", "q75", "min", "max"),
  inObservation = TRUE,
  standardConcept = TRUE,
  sourceVocabulary = FALSE,
  domainId = TRUE,
  typeConcept = TRUE
)

Arguments

omopTable

An omop_table object.

recordsPerPerson

Estimates to summarise the number of records per person.

inObservation

Whether to include the percentage of records in observation.

standardConcept

Whether to summarise standard concept.

sourceVocabulary

Whether to summarise source vocabulary.

domainId

Whether to summarise domain id of standard concept id.

typeConcept

Whether to summarise type concept id field.

Value

A summarised_result object with the summarised data.