Summarise a cdm_reference object creating a snapshot with the metadata of the cdm_reference object.

Summarise a cdm_reference object creating a snapshot with the metadata of the cdm_reference object.

Usage

summariseOmopSnapshot(cdm)

Arguments

cdm

A cdm_reference object.

Value

A summarised_result object that contains the OMOP CDM snapshot information.

Examples

# \donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")
#> ℹ Reading GiBleed tables.
#> ℹ Adding drug_strength table.
#> ℹ Creating local <cdm_reference> object.
#> ℹ Inserting <cdm_reference> into duckdb.

summariseOmopSnapshot(cdm = cdm)
#> # A tibble: 16 × 13
#>    result_id cdm_name group_name group_level strata_name strata_level
#>        <int> <chr>    <chr>      <chr>       <chr>       <chr>       
#>  1         1 GiBleed  overall    overall     overall     overall     
#>  2         1 GiBleed  overall    overall     overall     overall     
#>  3         1 GiBleed  overall    overall     overall     overall     
#>  4         1 GiBleed  overall    overall     overall     overall     
#>  5         1 GiBleed  overall    overall     overall     overall     
#>  6         1 GiBleed  overall    overall     overall     overall     
#>  7         1 GiBleed  overall    overall     overall     overall     
#>  8         1 GiBleed  overall    overall     overall     overall     
#>  9         1 GiBleed  overall    overall     overall     overall     
#> 10         1 GiBleed  overall    overall     overall     overall     
#> 11         1 GiBleed  overall    overall     overall     overall     
#> 12         1 GiBleed  overall    overall     overall     overall     
#> 13         1 GiBleed  overall    overall     overall     overall     
#> 14         1 GiBleed  overall    overall     overall     overall     
#> 15         1 GiBleed  overall    overall     overall     overall     
#> 16         1 GiBleed  overall    overall     overall     overall     
#> # ℹ 7 more variables: variable_name <chr>, variable_level <chr>,
#> #   estimate_name <chr>, estimate_type <chr>, estimate_value <chr>,
#> #   additional_name <chr>, additional_level <chr>
# }