Summarise a cdm_reference object creating a snapshot with the metadata of the cdm_reference object

Summarise a cdm_reference object creating a snapshot with the metadata of the cdm_reference object

Usage

summariseOmopSnapshot(cdm)

Arguments

cdm

A cdm_reference object. Use CDMConnector to create a reference to a database or omock to create a reference to synthetic data.

Value

A summarised_result object with the results.

Examples

# \donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")
#> ℹ Reading GiBleed tables.
#> ℹ Adding drug_strength table.
#> ℹ Creating local <cdm_reference> object.
#> ℹ Inserting <cdm_reference> into duckdb.

result <- summariseOmopSnapshot(cdm = cdm)

tableOmopSnapshot(result = result)

  
Snapshot of the cdm GiBleed
  
Estimate
      	
        
Database name

      
    

GiBleed
    

General
    

Snapshot date
	2025-12-04

Person count
	2,694

Vocabulary version
	v5.0 18-JAN-19

Cdm
    

Source name
	Synthea synthetic health database

Version
	5.3

Holder name
	OHDSI Community

Release date
	2019-05-25

Description
	SyntheaTM is a Synthetic Patient Population Simulator. The goal is to output synthetic, realistic (but not real), patient data and associated health records in a variety of formats.

Documentation reference
	https://synthetichealth.github.io/synthea/

Observation period
    

N
	2,694

Start date
	1908-09-22

End date
	2019-07-03

Cdm source
    

Type
	duckdb

Package
	CDMConnector

Write schema
	results

Write prefix
	test_

cdmDisconnect(cdm = cdm)
# }