Create a visual table from a summariseConceptIdCounts() result

Create a visual table from a summariseConceptIdCounts() result

Usage

tableConceptIdCounts(result, display = "overall", type = "reactable")

Arguments

result

A summarised_result object (output of summariseConceptIdCounts()).

display

A character string indicating which subset of the data to display. Options are:

"overall": Show all source and standard concepts.

"standard": Show only standard concepts.

"source": Show only source codes.

"missing standard": Show only source codes that are missing a mapped standard concept.

type

Type of formatting output table, either "reactable" or "datatable".

Value

A formatted table visualisation.

Examples

# \donttest{
library(OmopSketch)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")
#> ℹ Reading GiBleed tables.
#> ℹ Adding drug_strength table.
#> ℹ Creating local <cdm_reference> object.
#> ℹ Inserting <cdm_reference> into duckdb.

result <- summariseConceptIdCounts(cdm = cdm, omopTableName = "condition_occurrence")
tableConceptIdCounts(result = result, display = "standard")

cdmDisconnect(cdm = cdm)
# }