Visualise the results of summarisePerson() into a table

Visualise the results of summarisePerson() into a table

Visualise the output of summarisePerson()

Usage

tablePerson(result, style = "default", type = "gt")

tablePerson(result, style = "default", type = "gt")

Arguments

result

A summarised_result object.

style

A character string or custom R code to define the visual formatting of the table.

type

The desired format of the output table. See visOmopResults::tableType() for allowed options.

Value

A table visualisation.

A visualisation of the data summarising the person table.

Examples

# \donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summarisePerson(cdm = cdm)

tablePerson(result = result)

  

Variable name
      	Variable level
      	Estimate name
      	
        
CDM name

      
    

mockOmopSketch
    

Number subjects
	-
	N
	100

Number subjects not in observation
	-
	N (%)
	0 (0.00%)

Sex
	Female
	N (%)
	57 (57.00%)

	Male
	N (%)
	43 (43.00%)

	None
	N (%)
	0 (0.00%)

Sex source
	Missing
	N (%)
	100 (100.00%)

Race
	Missing
	N (%)
	100 (100.00%)

Race source
	Missing
	N (%)
	100 (100.00%)

Ethnicity
	Missing
	N (%)
	100 (100.00%)

Ethnicity source
	Missing
	N (%)
	100 (100.00%)

Year of birth
	-
	Missing (%)
	0 (0.00%)

	
	Median [Q25 - Q75]
	1,976 [1,958 - 1,989]

	
	90% Range [Q05 to Q95]
	1,951 to 1,997

	
	Range [min to max]
	1,950 to 2,000

Month of birth
	-
	Missing (%)
	0 (0.00%)

	
	Median [Q25 - Q75]
	6 [3 - 10]

	
	90% Range [Q05 to Q95]
	1 to 12

	
	Range [min to max]
	1 to 12

Day of birth
	-
	Missing (%)
	0 (0.00%)

	
	Median [Q25 - Q75]
	16 [8 - 24]

	
	90% Range [Q05 to Q95]
	2 to 30

	
	Range [min to max]
	1 to 31

Location
	-
	Missing (%)
	100 (100.00%)

	
	Zero count (%)
	0 (0.00%)

	
	Distinct values
	1

Provider
	-
	Missing (%)
	100 (100.00%)

	
	Zero count (%)
	0 (0.00%)

	
	Distinct values
	1

Care site
	-
	Missing (%)
	100 (100.00%)

	
	Zero count (%)
	0 (0.00%)

	
	Distinct values
	1

# }

# \donttest{
library(OmopSketch)

cdm <- mockOmopSketch(numberIndividuals = 100)

result <- summarisePerson(cdm = cdm)

tablePerson(result = result)

  

Variable name
      	Variable level
      	Estimate name
      	
        
CDM name

      
    

mockOmopSketch
    

Number subjects
	-
	N
	100

Number subjects not in observation
	-
	N (%)
	0 (0.00%)

Sex
	Female
	N (%)
	56 (56.00%)

	Male
	N (%)
	44 (44.00%)

	None
	N (%)
	0 (0.00%)

Sex source
	Missing
	N (%)
	100 (100.00%)

Race
	Missing
	N (%)
	100 (100.00%)

Race source
	Missing
	N (%)
	100 (100.00%)

Ethnicity
	Missing
	N (%)
	100 (100.00%)

Ethnicity source
	Missing
	N (%)
	100 (100.00%)

Year of birth
	-
	Missing (%)
	0 (0.00%)

	
	Median [Q25 - Q75]
	1,974 [1,959 - 1,987]

	
	90% Range [Q05 to Q95]
	1,953 to 1,999

	
	Range [min to max]
	1,951 to 2,000

Month of birth
	-
	Missing (%)
	0 (0.00%)

	
	Median [Q25 - Q75]
	7 [4 - 9]

	
	90% Range [Q05 to Q95]
	1 to 12

	
	Range [min to max]
	1 to 12

Day of birth
	-
	Missing (%)
	0 (0.00%)

	
	Median [Q25 - Q75]
	15 [8 - 23]

	
	90% Range [Q05 to Q95]
	2 to 29

	
	Range [min to max]
	1 to 31

Location
	-
	Missing (%)
	100 (100.00%)

	
	Zero count (%)
	0 (0.00%)

	
	Distinct values
	1

Provider
	-
	Missing (%)
	100 (100.00%)

	
	Zero count (%)
	0 (0.00%)

	
	Distinct values
	1

Care site
	-
	Missing (%)
	100 (100.00%)

	
	Zero count (%)
	0 (0.00%)

	
	Distinct values
	1

# }