Create a visual table from a summarise_population_characteristics result.

Create a visual table from a summarise_population_characteristics result.

Usage

tablePopulationCharacteristics(result, type = "gt")

Arguments

result

Output from summarisePopulationCharacteristics().

type

Type of formatting output table, either "gt" or "flextable".

Value

A gt or flextable object with the summarised data.

Examples

# \donttest{
cdm <- mockOmopSketch()

summarisedPopulation <- summarisePopulationCharacteristics(
  cdm = cdm,
  studyPeriod = c("2010-01-01", NA),
  sex = TRUE,
  ageGroup = list("<=60" = c(0, 60), ">60" = c(61, Inf))
)
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ℹ Building new trimmed cohort
#> Creating initial cohort
#> ✔ Cohort trimmed
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! The following column type were changed:
#> • variable_name: from integer to character

summarisedPopulation |>
  suppress(minCellCount = 5) |>
  tablePopulationCharacteristics()

  

Variable name
      	Variable level
      	Estimate name
      	
        
Database name

      
    

mockOmopSketch
    

overall; overall
    

Number records
	-
	N
	58

Number subjects
	-
	N
	58

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2014-10-28]

	
	Range
	2010-01-01 to 2019-11-25

Cohort end date
	-
	Median [Q25 - Q75]
	2015-01-23 [2013-01-31 - 2018-04-09]

	
	Range
	2010-04-14 to 2019-12-19

Age at start
	-
	Median [Q25 - Q75]
	31 [23 - 46]

	
	Mean (SD)
	34.14 (14.95)

	
	Range
	11 to 65

Age at end
	-
	Median [Q25 - Q75]
	36 [27 - 49]

	
	Range
	12 to 66

Sex
	Female
	N%
	31 (53.45)

	Male
	N%
	27 (46.55)

Prior observation
	-
	Median [Q25 - Q75]
	261 [0 - 2,740]

	
	Mean (SD)
	1,987.72 (3,039.53)

	
	Range
	0 to 12,363

Future observation
	-
	Median [Q25 - Q75]
	862 [368 - 1,542]

	
	Mean (SD)
	1,144.55 (1,004.22)

	
	Range
	4 to 3,631

Days in cohort
	-
	Median [Q25 - Q75]
	863 [370 - 1,543]

	
	Mean (SD)
	1,145.55 (1,004.22)

	
	Range
	5 to 3,632

<=60; overall
    

Number records
	-
	N
	56

Number subjects
	-
	N
	56

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2014-09-27]

	
	Range
	2010-01-01 to 2019-11-25

Cohort end date
	-
	Median [Q25 - Q75]
	2014-12-24 [2013-01-11 - 2018-03-08]

	
	Range
	2010-04-14 to 2019-12-19

Age at start
	-
	Median [Q25 - Q75]
	30 [23 - 45]

	
	Mean (SD)
	33.07 (14.07)

	
	Range
	11 to 58

Age at end
	-
	Median [Q25 - Q75]
	36 [26 - 48]

	
	Range
	12 to 62

Sex
	Female
	N%
	30 (53.57)

	Male
	N%
	26 (46.43)

Prior observation
	-
	Median [Q25 - Q75]
	368 [0 - 2,780]

	
	Mean (SD)
	2,058.71 (3,070.16)

	
	Range
	0 to 12,363

Future observation
	-
	Median [Q25 - Q75]
	888 [458 - 1,553]

	
	Mean (SD)
	1,173.70 (1,009.71)

	
	Range
	4 to 3,631

Days in cohort
	-
	Median [Q25 - Q75]
	890 [458 - 1,554]

	
	Mean (SD)
	1,174.70 (1,009.71)

	
	Range
	5 to 3,632

>60; overall
    

Number records
	-
	N
	<5

Number subjects
	-
	N
	<5

Cohort start date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Cohort end date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Age at start
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Age at end
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Sex
	Female
	N%
	<5 (<5)

	Male
	N%
	<5 (<5)

Prior observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Future observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Days in cohort
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

overall; Female
    

Number records
	-
	N
	31

Number subjects
	-
	N
	31

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2013-06-12]

	
	Range
	2010-01-01 to 2019-05-28

Cohort end date
	-
	Median [Q25 - Q75]
	2014-05-12 [2012-04-08 - 2018-05-11]

	
	Range
	2010-04-14 to 2019-12-19

Age at start
	-
	Median [Q25 - Q75]
	34 [24 - 46]

	
	Mean (SD)
	35.55 (13.76)

	
	Range
	13 to 63

Age at end
	-
	Median [Q25 - Q75]
	37 [28 - 50]

	
	Range
	15 to 64

Sex
	Female
	N%
	31 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	792 [0 - 3,968]

	
	Mean (SD)
	2,442.65 (3,180.65)

	
	Range
	0 to 9,470

Future observation
	-
	Median [Q25 - Q75]
	812 [497 - 1,559]

	
	Mean (SD)
	1,156.10 (970.74)

	
	Range
	71 to 3,595

Days in cohort
	-
	Median [Q25 - Q75]
	813 [498 - 1,560]

	
	Mean (SD)
	1,157.10 (970.74)

	
	Range
	72 to 3,596

overall; Male
    

Number records
	-
	N
	27

Number subjects
	-
	N
	27

Cohort start date
	-
	Median [Q25 - Q75]
	2011-12-16 [2010-01-01 - 2014-12-10]

	
	Range
	2010-01-01 to 2019-11-25

Cohort end date
	-
	Median [Q25 - Q75]
	2016-04-03 [2013-08-29 - 2018-01-31]

	
	Range
	2010-05-01 to 2019-12-17

Age at start
	-
	Median [Q25 - Q75]
	29 [18 - 46]

	
	Mean (SD)
	32.52 (16.33)

	
	Range
	11 to 65

Age at end
	-
	Median [Q25 - Q75]
	34 [21 - 48]

	
	Range
	12 to 66

Sex
	Male
	N%
	27 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	0 [0 - 2,322]

	
	Mean (SD)
	1,465.41 (2,837.47)

	
	Range
	0 to 12,363

Future observation
	-
	Median [Q25 - Q75]
	892 [285 - 1,456]

	
	Mean (SD)
	1,131.30 (1,059.80)

	
	Range
	4 to 3,631

Days in cohort
	-
	Median [Q25 - Q75]
	893 [286 - 1,457]

	
	Mean (SD)
	1,132.30 (1,059.80)

	
	Range
	5 to 3,632

<=60; Female
    

Number records
	-
	N
	30

Number subjects
	-
	N
	30

Cohort start date
	-
	Median [Q25 - Q75]
	2010-01-01 [2010-01-01 - 2011-11-06]

	
	Range
	2010-01-01 to 2019-05-28

Cohort end date
	-
	Median [Q25 - Q75]
	2014-05-02 [2012-04-02 - 2018-04-09]

	
	Range
	2010-04-14 to 2019-12-19

Age at start
	-
	Median [Q25 - Q75]
	33 [24 - 44]

	
	Mean (SD)
	34.63 (13.00)

	
	Range
	13 to 58

Age at end
	-
	Median [Q25 - Q75]
	36 [28 - 48]

	
	Range
	15 to 62

Sex
	Female
	N%
	30 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	826 [0 - 4,431]

	
	Mean (SD)
	2,524.07 (3,202.00)

	
	Range
	0 to 9,470

Future observation
	-
	Median [Q25 - Q75]
	826 [512 - 1,565]

	
	Mean (SD)
	1,187.67 (971.02)

	
	Range
	71 to 3,595

Days in cohort
	-
	Median [Q25 - Q75]
	826 [513 - 1,566]

	
	Mean (SD)
	1,188.67 (971.02)

	
	Range
	72 to 3,596

<=60; Male
    

Number records
	-
	N
	26

Number subjects
	-
	N
	26

Cohort start date
	-
	Median [Q25 - Q75]
	2011-12-01 [2010-01-01 - 2014-10-27]

	
	Range
	2010-01-01 to 2019-11-25

Cohort end date
	-
	Median [Q25 - Q75]
	2015-10-29 [2013-08-05 - 2017-12-06]

	
	Range
	2010-05-01 to 2019-12-17

Age at start
	-
	Median [Q25 - Q75]
	29 [18 - 46]

	
	Mean (SD)
	31.27 (15.28)

	
	Range
	11 to 58

Age at end
	-
	Median [Q25 - Q75]
	33 [20 - 47]

	
	Range
	12 to 60

Sex
	Male
	N%
	26 (100.00)

Prior observation
	-
	Median [Q25 - Q75]
	0 [0 - 2,422]

	
	Mean (SD)
	1,521.77 (2,878.20)

	
	Range
	0 to 12,363

Future observation
	-
	Median [Q25 - Q75]
	910 [280 - 1,491]

	
	Mean (SD)
	1,157.58 (1,071.78)

	
	Range
	4 to 3,631

Days in cohort
	-
	Median [Q25 - Q75]
	912 [281 - 1,492]

	
	Mean (SD)
	1,158.58 (1,071.78)

	
	Range
	5 to 3,632

>60; Female
    

Number records
	-
	N
	<5

Number subjects
	-
	N
	<5

Cohort start date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Cohort end date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Age at start
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Age at end
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Sex
	Female
	N%
	<5 (<5)

Prior observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Future observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Days in cohort
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

>60; Male
    

Number records
	-
	N
	<5

Number subjects
	-
	N
	<5

Cohort start date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Cohort end date
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Age at start
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Age at end
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Range
	<5 to <5

Sex
	Male
	N%
	<5 (<5)

Prior observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Future observation
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

Days in cohort
	-
	Median [Q25 - Q75]
	<5 [<5 - <5]

	
	Mean (SD)
	<5 (<5)

	
	Range
	<5 to <5

PatientProfiles::mockDisconnect(cdm = cdm)
# }