Create a gt table from a summarised population characteristics table.

Create a gt table from a summarised population characteristics table.

Usage

tablePopulationCharacteristics(result, type = "gt")

Arguments

result

Output from summarisePopulationCharacteristics().

type

Type of table.

Value

A visual table.

Examples

# \donttest{

# Connect to a mock database
cdm <- mockOmopSketch()

# Run summarise clinical tables
summarisedPopulation <- summarisePopulationCharacteristics(
  cdm = cdm,
  studyPeriod = c("2010-01-01",NA),
  sex = TRUE,
  ageGroup = list("<=60" = c(0, 60), ">60" = c(61, Inf))
)
#> Warning: ! 1 casted column in og_011_1727040170 (cohort_set) as do not match expected
#>   column type:
#> • `cohort_definition_id` from numeric to integer
#> Warning: ! 1 column in og_011_1727040170 do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ℹ Building new trimmed cohort
#> Warning: ! 1 column in tmp_006_og_013_1727040171 do not match expected column type:
#> • `cohort_definition_id` is numeric but expected integer
#> Creating initial cohort
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ✔ Cohort trimmed
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! The following column type were changed:
#> • variable_name: from integer to character

# Create a visual table
tablePopulationCharacteristics(summarisedPopulation)
#> ! Results have not been suppressed.

  

Variable name
      	Variable level
      	Estimate name
      	
        Database name
      
    

mockOmopSketch
    

overall; overall
    

Number records
	-
	N
	55

Number subjects
	-
	N
	55

Cohort start date
	-
	min
	2010-01-01

	
	q25
	2010-01-01

	
	median
	2010-01-01

	
	q75
	2015-10-16

	
	max
	2019-08-11

Cohort end date
	-
	min
	2010-07-15

	
	q25
	2015-03-01

	
	median
	2017-03-06

	
	q75
	2018-08-21

	
	max
	2019-11-18

Age at start
	-
	Mean (SD)
	34.02 (15.34)

	
	min
	9

	
	q25
	21

	
	median
	32

	
	q75
	46

	
	max
	63

Age at end
	-
	min
	11

	
	q25
	26

	
	median
	35

	
	q75
	50

	
	max
	66

Sex
	Female
	N%
	29 (52.73)

	Male
	N%
	26 (47.27)

Prior observation
	-
	Mean (SD)
	2,575.40 (4,228.53)

	
	min
	0

	
	q25
	0

	
	median
	592

	
	q75
	3,402

	
	max
	20,336

Future observation
	-
	Mean (SD)
	1,267.27 (1,104.82)

	
	min
	0

	
	q25
	304

	
	median
	762

	
	q75
	2,095

	
	max
	3,593

<=60; overall
    

Number records
	-
	N
	52

Number subjects
	-
	N
	52

Cohort start date
	-
	min
	2010-01-01

	
	q25
	2010-01-01

	
	median
	2010-01-01

	
	q75
	2014-12-06

	
	max
	2019-08-11

Cohort end date
	-
	min
	2010-07-15

	
	q25
	2014-10-13

	
	median
	2016-12-25

	
	q75
	2018-08-19

	
	max
	2019-11-18

Age at start
	-
	Mean (SD)
	32.38 (14.11)

	
	min
	9

	
	q25
	21

	
	median
	30

	
	q75
	44

	
	max
	59

Age at end
	-
	min
	11

	
	q25
	25

	
	median
	32

	
	q75
	46

	
	max
	62

Sex
	Female
	N%
	26 (50.00)

	Male
	N%
	26 (50.00)

Prior observation
	-
	Mean (SD)
	2,723.98 (4,303.44)

	
	min
	0

	
	q25
	0

	
	median
	830

	
	q75
	3,677

	
	max
	20,336

Future observation
	-
	Mean (SD)
	1,302.71 (1,122.65)

	
	min
	0

	
	q25
	308

	
	median
	821

	
	q75
	2,190

	
	max
	3,593

>60; overall
    

Number records
	-
	N
	3

Number subjects
	-
	N
	3

Cohort start date
	-
	min
	2013-06-01

	
	q25
	2015-01-14

	
	median
	2016-08-28

	
	q75
	2017-08-04

	
	max
	2018-07-11

Cohort end date
	-
	min
	2016-08-19

	
	q25
	2017-05-08

	
	median
	2018-01-25

	
	q75
	2018-08-31

	
	max
	2019-04-06

Age at start
	-
	Mean (SD)
	62.33 (1.15)

	
	min
	61

	
	q25
	62

	
	median
	63

	
	q75
	63

	
	max
	63

Age at end
	-
	min
	61

	
	q25
	62

	
	median
	64

	
	q75
	65

	
	max
	66

Sex
	Female
	N%
	3 (100.00)

Prior observation
	-
	Mean (SD)
	0.00 (0.00)

	
	min
	0

	
	q25
	0

	
	median
	0

	
	q75
	0

	
	max
	0

Future observation
	-
	Mean (SD)
	653.00 (468.50)

	
	min
	269

	
	q25
	392

	
	median
	515

	
	q75
	845

	
	max
	1,175

overall; Female
    

Number records
	-
	N
	29

Number subjects
	-
	N
	29

Cohort start date
	-
	min
	2010-01-01

	
	q25
	2010-01-01

	
	median
	2013-06-01

	
	q75
	2014-09-02

	
	max
	2019-04-18

Cohort end date
	-
	min
	2010-10-06

	
	q25
	2013-11-10

	
	median
	2016-08-19

	
	q75
	2018-12-13

	
	max
	2019-11-03

Age at start
	-
	Mean (SD)
	36.72 (16.52)

	
	min
	9

	
	q25
	23

	
	median
	34

	
	q75
	50

	
	max
	63

Age at end
	-
	min
	18

	
	q25
	29

	
	median
	35

	
	q75
	52

	
	max
	66

Sex
	Female
	N%
	29 (100.00)

Prior observation
	-
	Mean (SD)
	3,351.14 (5,227.00)

	
	min
	0

	
	q25
	0

	
	median
	0

	
	q75
	5,855

	
	max
	20,336

Future observation
	-
	Mean (SD)
	1,144.41 (1,093.75)

	
	min
	29

	
	q25
	278

	
	median
	570

	
	q75
	1,942

	
	max
	3,593

overall; Male
    

Number records
	-
	N
	26

Number subjects
	-
	N
	26

Cohort start date
	-
	min
	2010-01-01

	
	q25
	2010-01-01

	
	median
	2010-01-01

	
	q75
	2015-10-31

	
	max
	2019-08-11

Cohort end date
	-
	min
	2010-07-15

	
	q25
	2016-03-18

	
	median
	2017-04-10

	
	q75
	2018-04-11

	
	max
	2019-11-18

Age at start
	-
	Mean (SD)
	31.00 (13.58)

	
	min
	9

	
	q25
	20

	
	median
	26

	
	q75
	38

	
	max
	59

Age at end
	-
	min
	11

	
	q25
	26

	
	median
	32

	
	q75
	44

	
	max
	61

Sex
	Male
	N%
	26 (100.00)

Prior observation
	-
	Mean (SD)
	1,710.15 (2,558.31)

	
	min
	0

	
	q25
	0

	
	median
	743

	
	q75
	1,686

	
	max
	8,778

Future observation
	-
	Mean (SD)
	1,404.31 (1,122.34)

	
	min
	0

	
	q25
	446

	
	median
	1,046

	
	q75
	2,405

	
	max
	3,552

<=60; Female
    

Number records
	-
	N
	26

Number subjects
	-
	N
	26

Cohort start date
	-
	min
	2010-01-01

	
	q25
	2010-01-01

	
	median
	2010-01-01

	
	q75
	2014-07-10

	
	max
	2019-04-18

Cohort end date
	-
	min
	2010-10-06

	
	q25
	2013-08-23

	
	median
	2015-07-07

	
	q75
	2018-11-17

	
	max
	2019-11-03

Age at start
	-
	Mean (SD)
	33.77 (14.76)

	
	min
	9

	
	q25
	22

	
	median
	34

	
	q75
	45

	
	max
	59

Age at end
	-
	min
	18

	
	q25
	26

	
	median
	35

	
	q75
	49

	
	max
	62

Sex
	Female
	N%
	26 (100.00)

Prior observation
	-
	Mean (SD)
	3,737.81 (5,394.17)

	
	min
	0

	
	q25
	0

	
	median
	1,020

	
	q75
	5,901

	
	max
	20,336

Future observation
	-
	Mean (SD)
	1,201.12 (1,135.77)

	
	min
	29

	
	q25
	282

	
	median
	618

	
	q75
	1,961

	
	max
	3,593

<=60; Male
    

Number records
	-
	N
	26

Number subjects
	-
	N
	26

Cohort start date
	-
	min
	2010-01-01

	
	q25
	2010-01-01

	
	median
	2010-01-01

	
	q75
	2015-10-31

	
	max
	2019-08-11

Cohort end date
	-
	min
	2010-07-15

	
	q25
	2016-03-18

	
	median
	2017-04-10

	
	q75
	2018-04-11

	
	max
	2019-11-18

Age at start
	-
	Mean (SD)
	31.00 (13.58)

	
	min
	9

	
	q25
	20

	
	median
	26

	
	q75
	38

	
	max
	59

Age at end
	-
	min
	11

	
	q25
	26

	
	median
	32

	
	q75
	44

	
	max
	61

Sex
	Male
	N%
	26 (100.00)

Prior observation
	-
	Mean (SD)
	1,710.15 (2,558.31)

	
	min
	0

	
	q25
	0

	
	median
	743

	
	q75
	1,686

	
	max
	8,778

Future observation
	-
	Mean (SD)
	1,404.31 (1,122.34)

	
	min
	0

	
	q25
	446

	
	median
	1,046

	
	q75
	2,405

	
	max
	3,552

>60; Female
    

Number records
	-
	N
	3

Number subjects
	-
	N
	3

Cohort start date
	-
	min
	2013-06-01

	
	q25
	2015-01-14

	
	median
	2016-08-28

	
	q75
	2017-08-04

	
	max
	2018-07-11

Cohort end date
	-
	min
	2016-08-19

	
	q25
	2017-05-08

	
	median
	2018-01-25

	
	q75
	2018-08-31

	
	max
	2019-04-06

Age at start
	-
	Mean (SD)
	62.33 (1.15)

	
	min
	61

	
	q25
	62

	
	median
	63

	
	q75
	63

	
	max
	63

Age at end
	-
	min
	61

	
	q25
	62

	
	median
	64

	
	q75
	65

	
	max
	66

Sex
	Female
	N%
	3 (100.00)

Prior observation
	-
	Mean (SD)
	0.00 (0.00)

	
	min
	0

	
	q25
	0

	
	median
	0

	
	q75
	0

	
	max
	0

Future observation
	-
	Mean (SD)
	653.00 (468.50)

	
	min
	269

	
	q25
	392

	
	median
	515

	
	q75
	845

	
	max
	1,175

# delete mock data
PatientProfiles::mockDisconnect(cdm = cdm)
# }